Revert "Removed 'scalebar/utils' and fixed visibiity of `calculateEndingGlobalCoordinates`"
This reverts commit 49184ae898b63aa258df1a340585cab80c1f5d8f.
